Kill off LP64 warnings.

+++ contrib/programs/viewres/viewres.c
@@ -442,7 +442,7 @@ static void variable_labeltype_callback (gw, closure, data)
     XtPointer closure;			/* TRUE or FALSE */
     XtPointer data;
 {
-    set_labeltype_menu ((Boolean) closure, True);
+    set_labeltype_menu ((Boolean)(long) closure, True);
 }
 
 /* ARGSUSED */
@@ -451,7 +451,7 @@ static void gravity_callback (gw, closure, data)
     XtPointer closure;			/* TRUE or FALSE */
     XtPointer data;
 {
-    set_orientation_menu ((XtGravity) closure, True);
+    set_orientation_menu ((XtGravity)(long) closure, True);
 }
 
 
@@ -557,7 +557,7 @@ static void show_resources_callback (gw, closure, data)
     XtPointer closure;			/* BOOL_OFF, BOOL_ON, BOOL_TOGGLE */
     XtPointer data;			/* undefined */
 {
-    int op = (int) closure;
+    int op = (int)(long) closure;
     XmuWidgetNode *node = widget_to_node (gw);
 
     if (node) {
@@ -589,7 +589,7 @@ static void select_callback (gw, closure, data)
     int nselected = selected_list.n_elements;
     XmuWidgetNode *node;
 
-    switch ((int) closure) {
+    switch ((int)(long) closure) {
       case SELECT_NOTHING:		/* clear selection_list */
 	remove_nodes_from_selected_list (0, nselected, True);
 	break;
@@ -703,7 +703,7 @@ static void toggle_callback (gw, closure, data)
     XtPointer data;		/* on or off */
 {
     XmuWidgetNode *node = (XmuWidgetNode *) closure;
-    Boolean selected = (Boolean) data;
+    Boolean selected = (Boolean)(long) data;
 
     if (selected) {
 	add_to_selected_list (node, FALSE);
@@ -1188,7 +1188,7 @@ static void do_single_arg (w, params, nparams, table, nentries, proc)
     /*
      * use any old widget
      */
-    (*proc) (w, (XtPointer) obj, (XtPointer) NULL);
+    (*proc) (w, (XtPointer)(long) obj, (XtPointer) NULL);
 }
